+++ object_mother_pattern/models/base_mother.py
@@ -17,6 +17,16 @@ from typing import Any, Generic, Iterable, TypeVar, get_args, get_origin
 from uuid import UUID, uuid4
 
 from faker import Faker
+from faker.providers import user_agent
+
+try:
+    from value_object_pattern import ValueObject  # type: ignore[import-not-found]
+
+    HAS_VALUE_OBJECTS = True
+
+except ImportError:
+    HAS_VALUE_OBJECTS = False
+
 
 T = TypeVar('T')
 
@@ -51,6 +61,9 @@ class BaseMother(ABC, Generic[T]):  # noqa: UP046
                 if not isclass(object=mother_type) and get_origin(tp=mother_type) is None:
                     raise TypeError(f'BaseMother[...] <<<{mother_type}>>> must be a type. Got <<<{type(mother_type).__name__}>>> type.')  # noqa: E501  # fmt: skip
 
+                if HAS_VALUE_OBJECTS and isclass(object=mother_type) and issubclass(mother_type, ValueObject):
+                    mother_type = mother_type.type()
+
                 cls._type = mother_type
                 return
 
@@ -64,7 +77,10 @@ class BaseMother(ABC, Generic[T]):  # noqa: UP046
         Returns:
             Faker: Faker instance.
         """
-        return Faker(use_weighting=False)
+        faker = Faker(use_weighting=False)
+        faker.add_provider(user_agent)
+
+        return faker
 
     @classmethod
     @abstractmethod