+++ tests/mothers/identifiers/uuid/test_string_uuid_mother.py
@@ -2,10 +2,12 @@
 Test module for the StringUuidMother class.
 """
 
+from random import choice, choices
 from uuid import UUID
 
 from pytest import mark, raises as assert_raises
 
+from object_mother_pattern.mothers import IntegerMother
 from object_mother_pattern.mothers.identifiers import StringUuidMother
 
 
@@ -17,7 +19,7 @@ def test_string_uuid_mother_happy_path() -> None:
     value = StringUuidMother.create()
 
     assert type(value) is str
-    UUID(value)
+    assert UUID(value).version in {1, 3, 4, 5}
 
 
 @mark.unit_testing
@@ -59,3 +61,41 @@ def test_string_uuid_mother_invalid_value_type() -> None:
         match=r'StringUuidMother value must be a string.',
     ):
         StringUuidMother.create(value=StringUuidMother.invalid_type())
+
+
+@mark.unit_testing
+def test_string_uuid_mother_allow_versions() -> None:
+    """
+    Test StringUuidMother create method with allow_versions parameter.
+    """
+    version = choice(seq=(1, 3, 4, 5))  # noqa: S311
+    value = StringUuidMother.create(allow_versions={version})
+
+    assert type(value) is str
+    assert UUID(value).version == version
+
+
+@mark.unit_testing
+def test_string_uuid_mother_invalid_allow_versions_type() -> None:
+    """
+    Test StringUuidMother create method with invalid allow_versions type.
+    """
+    with assert_raises(
+        expected_exception=TypeError,
+        match='UuidMother allow_versions must be a set',
+    ):
+        StringUuidMother.create(allow_versions=[1, 2, 3])  # type: ignore[arg-type]
+
+
+@mark.unit_testing
+def test_string_uuid_mother_invalid_allow_versions_value() -> None:
+    """
+    Test StringUuidMother create method with invalid allow_versions value.
+    """
+    versions = choices(population=(10, 14, 7653, 13456), k=IntegerMother.positive())  # noqa: S311
+
+    with assert_raises(
+        expected_exception=ValueError,
+        match=r'UuidMother allow_versions must be a subset of \{1, 3, 4, 5\}',
+    ):
+        StringUuidMother.create(allow_versions=set(versions))